dahdi-base: test: Serialize ioctl_echocancel
We were able to trigger oops in the lab caused by two threads in the same
process simultaneously destroying and accessing the ops struct through calls to
ioctl_echocancel. It's unclear how best to protect this section, this patch
takes a heavy handed approach by locking the entire echocan create/destroy
ioctl.

This patch is intended as a field test to help a customer who may or may not be
experiencing a similar issue due to glare.

Signed-off-by: Russ Meyerriecks <rmeyerriecks@digium.com>